+++ resources/js/common/layouts/AuthBase.vue
@@ -1,42 +1,48 @@
 <script setup lang="ts">
+  import {Head} from '@inertiajs/vue3';
   import useCraftData from '@/common/composables/useCraftData';
   import craftCmsLogoUrl from '@public/images/craftcms.svg';
+  import {t} from '@craftcms/cp';
 
   const {general, system} = useCraftData();
 </script>
 
 <template>
-  <div class="cp-login">
-    <div class="grid gap-3 justify-items-center">
-      <h1 class="flex justify-center" v-if="general.cpLogoUrl">
+  <Head :title="t('Sign In')"></Head>
+  <main class="cp-login">
+    <div class="cp-login__wrapper grid gap-3 justify-items-center">
+      <h1 class="flex justify-center">
         <img
+          v-if="general.cpLogoUrl"
           :src="general.cpLogoUrl"
           :alt="system.name"
           class="cp-login__logo"
           width="288px"
         />
+        <template v-else>
+          {{ system.name }}
+        </template>
       </h1>
 
-      <div class="w-sm">
+      <div class="cp-login__form-wrapper">
         <slot></slot>
       </div>
 
       <a
         class="cp-login__powered-by"
-        href="http://craftcms.com/"
+        href="https://craftcms.com/"
         title="Powered by Craft CMS"
-        aria-label="Craft CMS"
       >
         <img
           :src="craftCmsLogoUrl"
-          alt=""
+          alt="Craft CMS"
           class="cp-login__craft-logo"
           width="104"
           height="26"
         />
       </a>
     </div>
-  </div>
+  </main>
 </template>
 
 <style scoped lang="scss">
@@ -46,12 +52,22 @@
     height: 100vh;
     place-items: center;
     gap: var(--c-spacing-md);
+    padding: var(--c-spacing-lg);
+  }
+
+  .cp-login__wrapper {
+    width: 100%;
+    max-width: var(--container-sm);
   }
 
   .cp-login__logo {
     display: inline-block;
     width: calc(288rem / 16);
     height: auto;
+  }
+
+  .cp-login__form-wrapper {
+    width: 100%;
   }
 
   .cp-login__powered-by {